+++ docs/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.4.0] - 2022-04-05
+
 ### Added
 
 - an option to write filter scores to a file with `opusfilter-test`
@@ -120,7 +122,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 First tagged version.
 
 
-[Unreleased]: https://github.com/Helsinki-NLP/OpusFilter/compare/2.3.1...develop
+[Unreleased]: https://github.com/Helsinki-NLP/OpusFilter/compare/2.4.0...develop
+[2.4.0]: https://github.com/Helsinki-NLP/OpusFilter/compare/2.3.1...2.4.0
 [2.3.1]: https://github.com/Helsinki-NLP/OpusFilter/compare/2.3.0...2.3.1
 [2.3.0]: https://github.com/Helsinki-NLP/OpusFilter/compare/2.2.0...2.3.0
 [2.2.0]: https://github.com/Helsinki-NLP/OpusFilter/compare/2.1.2...2.2.0